+++ extensions/control-plane/edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/controller/dto/auth/BearerAuthDTO.java
@@ -39,8 +39,6 @@ public record BearerAuthDTO(URI uri, String clientId, String clientSecret, Strin
     public BearerAuthDTO {
         Objects.requireNonNull(clientId);
         Objects.requireNonNull(clientSecret);
-        Objects.requireNonNull(username);
-        Objects.requireNonNull(password);
     }