+++ system-tests/standalone/test.sh
@@ -5,6 +5,10 @@ set -euo pipefail
 # Load utility functions
 source "system-tests/util.sh"
 
+# Step -1: Use in-memory-vault, functionally the test will be the same.
+cp ./launchers/standalone/build.gradle.kts ./launchers/standalone/build.gradle.kts.bak
+sed -i '/implementation(libs\.edc\.vault\.hashicorp)/d' ./launchers/standalone/build.gradle.kts
+
 # Step 0: Boot provider EDC (without AAS extension, this is the one we attach to)
 echo "Starting control-plane (using consumer launcher as image)"
 start_runtime consumer control-plane.properties
@@ -98,3 +102,9 @@ should_be_data=$(< system-tests/resources/aas.json jq -S -r '."conceptDescriptio
 diff -w <(echo "$received_data") <(echo "$should_be_data") > /dev/null
 
 echo "Transferred data checks out. Test complete!"
+
+
+# Step -1: Use in-memory-vault, functionally the test will be the same.
+echo "Reinstating hashicorp vault dependency in standalone build file"
+cp ./launchers/standalone/build.gradle.kts.bak ./launchers/standalone/build.gradle.kts
+rm ./launchers/standalone/build.gradle.kts.bak